+++ tests/integration/test_modelsearch.py
@@ -1,7 +1,7 @@
 import shutil
 
 from pharmpy import Model
-from pharmpy.modeling import first_order_absorption, set_first_order_elimination
+from pharmpy.modeling import set_first_order_absorption, set_first_order_elimination
 from pharmpy.tools.modelsearch import ModelSearch
 from pharmpy.utils import TemporaryDirectoryChanger
 
@@ -13,7 +13,7 @@ def test_exhaustive_stepwise(tmp_path, testdata):
         model = Model('mox2.mod')
         model.dataset_path = tmp_path / 'mx19B_ZO.csv'
 
-        first_order_absorption(model)
+        set_first_order_absorption(model)
         set_first_order_elimination(model)
         model.update_source(force=True)