+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ next version
 * Rename bolus_absorption to set_bolus_absorption
 * Rename seq_zo_fo_absorption to set_seq_zo_fo_absorption
 * Rename have_zero_order_absorption to has_zero_order_absorption
+* Rename power_on_ruv to set_power_on_ruv
 
 0.28.0 (2021-08-24)
 -------------------